Modified over GOF and added wc_max to calibration.
The GOF is now computed by combining the mNSE for Q and the pBias for
ETR. The mNSE is reduced multiplicatively using the pBias.

Since wc_max can have a strong impact on ETR, it was added as a
calibrated parameter to the range table and the update template. In that
course, thr_surf and fac_precip were removed from calibration because
they are kept constant anyway.